[XLA:GPU] Implement i4 support as a Triton IR rewrite.
Right now the triton emitter for multiply emits the code that operates with the i4 tensors packed into i8 with one 2x smaller dimension together with the unpacking steps. It makes sense to rework this taking into the account the fact that we also want to replace these emitters with the very complicated tailing logic with the new triton emitters.

The emitter could generate the code that operates with i4 tensors as is. I.e. emit the ops with AxBxi4 tensors and use ExtSI when we need to get i8. That would make the emitter simpler. After that we could do a Triton IR rewrite pass that would convert these i4 ops to i4 packed into i8 ops, and replace ExtSI to the unpacking sequence.

The cl is the example of such rewriter.

PiperOrigin-RevId: 705475032