Add a layout field to `xla::ifrt::ArraySpec`
This is the most natural way to incrementally add layout support. `layout` can be nullptr, in which case it indicates a default layout, or non-existent layout info if the runtime and/or DType do not support layouts.

For now, the layout field uses `std::shared_ptr<const xla::PjRtLayout>` as the type, but this can be switched to an IFRT-native layout type once it's introduced. The serialization assumes that the layout is always `xla::PjRtXlaLayout`, which holds true because that is the only available `xla::PjRtLayout` implementation today. This isn't ideal, but it is the only feasible way due to the lack of type-erased deserialization for `xla::PjRtLayout`. We don't have to worry about the backward compatibility of the serialization format right now because IFRT Proxy does not support layouts yet.

PiperOrigin-RevId: 712384508