Disable Embedding Pipelining When Recording Summaries
Embedding pipelining requires a least two steps. However, because the inclusion of summary ops is expensive and only needed periodically, most users run a single training step when enabling summaries. This change detects when summaries are active and automatically disables pipelining (under the assumption that the user will only be running a single step).

PiperOrigin-RevId: 716804368