Adds a library to make fixed options flags easier and safer to define.
xla defines custom commandline flag parsing logic in several places. These
definitions follow a common pattern: there is a fixed set of options for the
flag type (e.g. the flag type is an enum).

Such definitions are repetitious and error-prone (one has to make sure that the name => value and value => name mappings are consistent and have no duplicate names or values). This library abstracts away the tedious details and will be used in subsequent changes to make the definitions simpler and safer.

PiperOrigin-RevId: 723557007