Fix GatherClientLibraryTest under PjRt.
It was unclear to me what this test was actually trying to achieve because it is
not clearly documented. If it is trying to exercise a specific behavior with the
old non-PjRt `Client`, then that should probably live elsewhere. I've converted
it to something that can just run on top of `HloPjRtTestBase` directly. I added
some boilerplate to allow the `XlaBuilder` code to remain, though it might be
better to just use a HLO string directly.

PiperOrigin-RevId: 723677583