[PJRT:CPU] Derive ExecutableRunOptions::run_id from ExecuteOptions::launch_id
This change uses `ExecuteOptions::launch_id` (> 0) as
`ExecutableRunOptions::run_id`. This allows using a consistent `run_id` to be
used for a set of executions dispatched using per-device calls of
`PjRtLoadedExecutable::ExecuteSharded()`, matching the capability of the
sharded executions on TPU.

If `launch_id` is unset (= 0), it falls back to the previous behavior of using
an intenral counter that is incremented for any execution.

JAX executables now uses an increasing sequence of `launch_id` for each
execution. Its initial value is set to the fingerprint of the executables to
preserve the existing capability of choosing low-collision `launch_id` between
different executables.

PiperOrigin-RevId: 726258012