litert: Handle tensors are shared with CPU kernels
Currently, it uses NPU requirements When a tensor is shared with bot CPU and NPU.
This CL changes to return a CPU friendly requirements to make it runnable with CPU kernels.
In NPU, it will creates own NPU TensorBuffer and copy the content.

To check the shared tensors, introduced the CheckCpuTensors() method and cpu_tensors_ set.

Also updated SetCustomAllocationForTensor() to compatible with kTfLiteNonCpu type.

PiperOrigin-RevId: 728300176