Fold mhlo.reduce when the reduce body is empty.
This change targets reduce ops with an empty body and constant inputs. In such
a scenario, the reduce operation is simply creating a new constant with the
initial values provided by the second argument.

This new function is very similar in nature to `tryFoldOutsideValuesReduction`,
with the difference being that we go a step further in the case that the return
value is a constant.

PiperOrigin-RevId: 734200775