[xla:copy_insertion] Fix a problem in handling Send feeding into a while-loop.
Previously, we only make a copy of the operand for rotated Send/SendDone inside
a while-loop. We now also make a copy of the operand for the Send feeding into
a while-loop so that the buffer for the operand does not have to hold different
values with live range interference.

PiperOrigin-RevId: 736527447