Introduce registry for GPU helper kernels
We have a bunch of helper kernels that are getting called from various parts
of the compiler and the runtime (MakeBatchPointers, RedzoneAllocator, RepeatBuffers, etc.).

They all need to be implemented separately for CUDA and ROCm and makes the code depend
CUDA and ROCm specifics.

This code introduces a global registry for these kernels. Users of these kernels can
load a kernel via a trait type which also makes loading and registering a kernel type-safe.

PiperOrigin-RevId: 737892624