[XLA:GPU] Allow symbolic tile analysis to go through `concatenate`s for nested GEMM fusions.
We only enable a restricted case of concatenations for now, where tiles never
propagate to more than a single operand. Concretely, we enforce the following
constraints:

1. the start `offset` along the concatenation dimension is `0`;
2. the `stride` is always `1` along the concatenation dimension;
3. the `size` is a divisor of the size of the concatenation dimension of each
operand.

These constraints are slightly over restrictive (starting at a non-`0` offset
that is a multiple of the tile size would also be fine, and so would be
having a non-`1` stride so long as `stride * size` is a divisor of the size of
the concatenation dimension of each operand). Nevertheless, they're most
likely sufficient to complete the migration of GEMMs to the generic Triton
emitter.

PiperOrigin-RevId: 739119566