Refactor `hlo_memory_scheduler`.
Convert the `MemorySchedulerAlgorithm` and `ModuleSchedulerAlgorithm` type aliases into proper classes with inheritance relationship. This removes the need for the `ComputationSchedulerToModuleScheduler` function.

This design scheme is simpler and easier to extend/modify.

Also rename `PostOrderMemoryScheduler` to `PostOrderScheduler` and `BFSMemoryScheduler` to `BFScheduler`. These two variants don't minimise memory usage which makes the original names a bit misleading.

PiperOrigin-RevId: 739733653