Changes a misleading function name and reduces its usage.
`protobuf_util::ProtobufEquals(p1, p2)` sounds like comparing two protos for equality, but in reality it's comparing the serialized strings of the protos. Proto serialization isn't guaranteed to be stable, and may contain fields we don't care about in case the proto definition is subsequently extended. Therefore its behavior is dangerous and should be used only when we intend to compare serialized strings.

Rename it to `HaveSameSerialization()` to avoid misleading the readers.

Also replaces its uses in tests with `EqualsProto()`, which has a deterministic semantics and generates far better failure messages.

PiperOrigin-RevId: 740418850