[XLA:GPU] Round the lowest part of f32 instead of masking it for BF16_Xn algorithms
The rounding increases the precision of BF16_X3.
Relative error decreases from 2e-5 to 7e-6.

Initially we zeroed the last 16 bits of f32 for high part and for low part.
Now we zero the high part and round the low part.

The we do the same for X6 and X9 - zero high and med parts and round low part.

PiperOrigin-RevId: 741451522