[PjRt-GPU] Populate PjRt GPU client device's `process_index` property for multi-host
`xla::PjRtStreamExecutorDeviceDescription` takes a `process_index` to give a
correct `process_index` property. This information is not fully propagated in a
few places that all `xla::PjRtStreamExecutorDeviceDescription` returns
`process_index() == 0` regardless of their owner processes.

This CL makes a minimal change to populate
``xla::PjRtStreamExecutorDeviceDescription::process_index`, either by
forwarding the owner client's process index, or computing it from global device
ids that a device description represents.

This CL also removes the default value for `process_index` in both device and
device description constructors. This is to help propagate `process_index` in
the future code without accidentally assuming it to be 0.

PiperOrigin-RevId: 745369373