Fix undefined behaviors converting between `tpu::TpuTopology*` and `SE_TpuTopology*`.
`tpu::TpuTopology` and `SE_TpuTopology` are unrelated types. Therefore, the C++ compiler is allowed to assume that pointers to these two types can *never* be aliases, and `reinterpret_cast` between these pointers are undefined behavior.

The C++ strict aliasing rule makes an exception for pointers to byte-like types, e.g. `char*`. It allows a `char*` to be an alias of a `tpu::TpuTopology*`.

To fix such undefined behaviors, we introduce an `SE_TpuTopologyPtr` struct that holds a `const char*` as an alias to a `const tpu::TpuTology*`, and use `SE_TpuTopologyPtr` where `SE_TpuTopology*` was used.

This change also improves the readability/safety of the code by removing many uses of `reinterpret_cast` and `const_cast`.

Also fix violation of

PiperOrigin-RevId: 747479267