[XLA] Fix bug in collective permute HLO verifier.
Collective permute has different semantics for operand depending if the instruction is inplace or not. Inplace collective permute always has 4 operands  and operand 2 and 3 are always tuples. Current code assumes that any collective permute with 4 operands is inplace, but that's not correct. If a not inplace collective permute has 4 operands that are not tuples, it causes a crash.

PiperOrigin-RevId: 757721062