[XLA:GPU][SymbolicTileAnalysis] Refactor the analysis to derive tiles with regards to all the tiling parameters in the fusion.
Previously, `SymbolicTile`s were derived based on the output tile sizes of
their direct parent computation. This made it complicated to aggregate
constraints consistently, since dimension parameters ended up having
computation-dependent meaning. E.g., take the following pseudo-HLO:

```
lhs_computation {
  p0 = ...
  ROOT reshape = f32[k,m] reshape(p0)
}

fusion_computation {
  p0 = ...
  p1 = ...
  lhs = f32[k,m] fusion(p0), calls=lhs_computation, backend_config=...
  rhs = f32[n,k] fusion(p1), ...
  ROOT dot = f32[m,n] dot(lhs, rhs), lhs_contracting_dimensions={0}, ...
}
```

In `fusion_computation`, we were working with indexing maps of the form
`(d0, d1) -> ...`, where `d0` and `d1` corresponded to the tile sizes at the
output of the fusion (`m`, and `n`).

But in `lhs_computation`, we were working with indexing maps of the form
`(d0, d1) -> ...`, where `d0` and `d1` corresponded to the tile sizes at the
output of the **`lhs`** fusion (`k` and `m`).

Due to this inconsistency, the constraints derived on the `reshape` instruction
were defined in terms of the tile sizes at the output of `lhs` (`k` and `m`),
but were checked against instantiations of `m` and `n` when
calling `ParametersSatisfyConstraints` later.

In the new version of the code, the maps always take as many parameters as
required for the whole fusion, which ensures that parameters always have the
same meaning throughout the entire fusion. In this example, the maps inside
both `fusion_computation` and `lhs_computation` are now of the form
`(d0, d1, d2) -> ...`, where `d0` has meaning `k`, `d1` meaning `m`, and
`d2` meaning `n`.

PiperOrigin-RevId: 763391149