[XLA:GPU] Remove the check that operands are not transposes in cuBLAS rewrite
Initially it was introduced as an assert, because DotNormalizer would supposedly fold the transposes into the dot dimension numbers.

Then we got LayoutAssignment which does insert transposes, especially for fp8 dots, so checks for invariants were replaced with a condition to decide whether to rewrite. (wrongly, because cuBLAS is fine to rewrite, it was just a check for invariant which stopped to hold).

This check doesn't bring any benefit anymore and should be removed.

The concrete case where this check became problematic is due to another issue which I'll fix in a follow up PR:
- We attempt to rewrite a dot with cuBLASlt, then with cuBLAS, in two separate passes.
- The passes assume that the decision whether to rewrite in principle would be identical, just the decision whether to rewrite the dot with cuBLASlt vs cuBLAS could be different.
- First, cuBLASlt rewrite pass "thinks" that it won't rewrite the dot because of Tranpose operand, but tweaks the computation so that operand is no longer transpose
- Second, cuBLAS rewrite "thinks" that it's fine to rewrite fp8 GEMM (which should either not be rewritten at all, or rewritten through cuBLASlt)
- Then cuBLAS call crashes as it doesn't know how to handle fp8

PiperOrigin-RevId: 767073663