[XLA:GPU]: Use double buffering for the one-shot all-reduce kernel
Our current implementation of the all-reduce uses synchronization twice
(beginning and end). We do this to make sure that a second invocation of
all-reduce on a different GPU does not overwrite memory still being used by the
first iteration.

Using a double buffer eliminates the need for the double synchronization.

PiperOrigin-RevId: 768027494