[XLA] Allow ReshapeMover to perform chain moves.
Moving a reshape down may create a chain of new opportunities. Previously, exploiting this required a pipeline to run ReshapeMover to a fixed point, but this is wasteful in many cases, since we can check whether an opportunity was created an exploit it immediately.

This can eliminate excessive iteration in fixed-point HLO pipelines that have both reshape-mover and algsimp.

PiperOrigin-RevId: 768094789