Make GpuTestKernels use the GpuKernelRegistry
This is removing the last user of `gpu_kernel_library` - which are the kernels that we have for testing. They are now also available through the `GpuKernelRegistry` and are getting linked automatically through the `xla_test` macro - that is ROCm tests link the ROCm version and CUDA tests link the CUDA version.

PiperOrigin-RevId: 769480027