Change `SerDesVersion`'s enum to `SerDesVersionNumber`
It has been identified that using enum to encode version number literals for
`SerDesVersion` causes a maintenance problem: as long as there is any SerDes
implementation that uses an enum to represent a certain format version, that
enum cannot be removed however old the enum is. For example, many IFRT types
would use `SerDesVersion::kV0Initial` as their initial format. If there is a
single IFRT type that keeps using the same format, we would not be able to
remove `SerDesVersion::kV0Initial`.

Among a couple of mitigation strategies, we choose to remove enums and use version number
literals directly in the SerDes code. This eliminates a dependency from SerDes
implementations to SerDes global version number literals, avoiding the above
problem. `SerDesVersionNumber` is a strongly typed integer, which prevents
implicit conversion from/to other integer types and slightly improve
searchability.

This change also adds instructions for how to add new formats and remove old
formats.

PiperOrigin-RevId: 770401696