[XLA:GPU][SymbolicTile] Allow `SymbolicTile` to not bail out when deriving from an `IndexingMap` with "right-padding" constraints.
We do intend to support the case where tile sizes are larger than the
restrictions on the input space, and therefore claim that this should be safe.
(And if it turns out not to be, then it'll be symptomatic of a bug.)

Modelling masked propagation more explicitly is a topic for the ongoing
rewrite of `SymbolicTile`. Unfortunately, since this infrastructure is not
yet available, we instead fix here.

PiperOrigin-RevId: 772833410