[XLA] Don't try to simplify While ops if their body or condition has more than one use.
This isn't something we currently expect to see in practice, but it could happen (e.g. if we make a shallow copy of an outlined call body). So while this is possible to handle, there's no point in doing the work right now. However, it's prudent to bail instead of trying to perform an invalid transform.

PiperOrigin-RevId: 773144308