[xla:gpu] Nest gemm fusion: only hoist bitcasts upwards and support layouts.
Instead of hoisting bitcasts downwards, insert a root bitcast which cancels out all bitcasts between the dot and the root.

This allows removing epilogue bitcasts in the presence of binary elementwise ops (which did not work before because the consumer set is not closed).

Support non-default layouts when hoisting bitcasts across transposes and broadcasts.

PiperOrigin-RevId: 774739343