[IFRT] Facilitate nested SerDes versioning with a bad default SerDes version
This change adds `SerDesDefaultVersionAccessor::Get()`, which returns the
default `SerDesVersion` (=`SerDesVersion::current()` normally) for any default
value for `SerializeOptions::version` and `ToProto()` `version` argument.

When `IFRT_TESTING_BAD_DEFAULT_SERDES_VERSION` macro is defined, this default
version is set to `-1`. This will cause serialization with no explicit version
to fail (assuming the serialization does support versioning). This is
particularly helpful for IFRT Proxy when ensuring that nested serialization is
plumbing the SerDes version correctly as every serialization in IFRT Proxy is
expected to use an explicit version that is negotiated between the Proxy client
and server.

PiperOrigin-RevId: 774984741