[IFRT] Rename `Client::GetDefaultLayout()` to `Client::GetDefaultPjRtLayout()`
`xla::ifrt::Client::GetDefaultLayout()` will be migrated to use
`xla::ifrt::CustomLayoutRef`, with a change to its arguments to take a global
(not shard) shape and sharding, which will make it more streamlined for
supporting memory-kind-specific default layouts and remove the need for manual
shard shape calculation at user callsites. For incremental migration of the
early changes, we will take similar steps to `Array::layout()` migration to
rename the method and add redirection.

PiperOrigin-RevId: 776718046