[XLA:GPU] [NFC] Move Algebraic Simplifier configs to one place
It was hard to track what config runs when, and when you want to enable and disable some transformation, it's hard to find all places you have to touch.

The change is intended to be a NOOP change. The attempt to actually make configs more unified will come in a later change.

PiperOrigin-RevId: 780965165