Use a literal sentinel value for kernel init failure
If a TFLite delegate is compiled into a library then `TfLiteKernelInitFailed` can return a different value than the one that the TFLite interpreter expects. This could be fixed by modifying the linkage of this function but that doesn't seem to be common in TFLite so the easier solution is to use a literal sentinel value (-1) instead of a memory address.

PiperOrigin-RevId: 784453957