Avoid crash in RequestActivityBuffer
Activity tracing is currently not getting disabled for the CONCURRENT_KERNEL type due to a deadlock issue in CUDA 12.8+.

This can lead to crashes during kernel execution after tracing has been disabled and the activity buffers have been destructed. The callbacks are still in place and will try to access the activity buffers during the next kernel execution.

To avoid these crashes I'm adding this check for the existence of the activity buffers.

PiperOrigin-RevId: 785411975