Make RedzoneBuffer constructible from ProgramShape
This adds a new factory function to the `RedzoneBuffers` type. It now can be constructed from a `ProgramShape` value.

I'm also changing the implementation of the other two factory function which
can now be implemented in terms of the new FromProgramShape factory function.

I also had to change the parameter to `CreateInputs` and `CreateOutputs` which used to operate on HloInstructions, but all they need are shapes.

This will help making the autotuner independent of HLO since the ProgramShape can be requested from a GpuExecutable without accessing the HLO module.

PiperOrigin-RevId: 799909951