Fix HloValueAnalysis to work with the latest attribute  _xla_host_transfer_rendezvous.
Previously, the pass expects a Send/Recv pair to have the same string value in attribute _xla_host_transfer_rendezvous, this is not always true. We have to remove the substrings args_dtoh_ or retvals_htod_ from the attribute value, if there are such subtrings, to find the matching Send/Recv pair.

PiperOrigin-RevId: 810097899